requirements.txt
streamlit>=1.41.0
google-genai>=1.0.0
openai>=1.50.0
python-dotenv>=1.0.0
